Rewrote lexer
Due to the one token lookahead, we might be parsing a token in the wrong mode.
E.g. \text{sin}^2 will parse the ^ in text mode before it gets a chance to
close its argument group and re-lex that position in math mode.
For this reason it is desirable if the set of valid text and math tokens
is the same, or if at least every valid text token starts with a valid math
token and vice versa.
In this new implementation, all the various things that could get matched
have been collected into a single regular expression.  The hope is that
this will be beneficial for performance and keep the code simpler.
The code was written with Unicode input in mind, including non-BMP codepoints.