Experiment with some font checking
Might be that the "check.py" should go to the test directory eventually, but
for now I'll leave it here until we have decided what to do about it.

The script checks various things about the fonts for consistency.
In particular, it verifies that unicode to glyph mappings are consistent
and checks whether all symbols in the fonts have associated metrics.
The latter is not true yet. Furthermore, it points out situations
where the glyph name in the FFT used by KaTeX does not agree with
the corresponding glyph name in the PFB used by e.g. dvips.
Many of these naming inconsistencies are harmless, but they should be
manually checked. Furthermore, any change to the list of such mismatches
should be detected and reported in the future, if we automate these tests.

As a side effect, the script creates one HTML file for each TTF font,
giving a chart of all the characters included in that font, with unicode
code point, KaTeX font rendering, rendering using the browser's default
font, and various names taken from the TTF file, the PostScript font
file which TeX uses, and the unicode database.  The name of the TeX font
is indicated as well.