Upload current KaTeX installation and test cases
The goal is to have publicly available versions of the test cases which can
be used to create screenshots using BrowserStack.  The test cases are stored
in separate HTML files, not transmitted via query strings, in order to avoid
cross-site scripting attacks on github io pages by malicious HTML code in
the pre or post settings.