Work around screenshotter getting stuck
For reasons unknown, the screenshotter may get stuck on Travis builds (see
issue #499).  It appears to do so after building the driver but before the
first screenshot for size adjustments has arrived.  To work around this
problem, we start a watchdog timer which will kill the node process if there
is no call to the heartbeat function in a certain (configurable) interval.

This also fixes a bug where failure to start a docker process would not lead
to a non-zero return value from screenshotter.sh.